+++ app/components/dataSet/DataSetValueActions.js
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 import {bindActionCreators} from 'redux';
 import {connect} from 'react-redux';
 import {Link, hashHistory} from 'react-router';
-import {Col, Button, Select, Menu, Dropdown, Icon} from 'antd';
+import {Col, Button, Select, Menu, Dropdown, Icon, Modal} from 'antd';
 import * as dataSetActions from '../../actions/dataSet';
 import * as app from '../../constants/app';
 
@@ -11,7 +11,7 @@ import * as app from '../../constants/app';
 const DATASET_DISABLE = 'DATASET_DISABLE';
 const DATASET_ENABLE = 'DATASET_ENABLE';
 const DATASET_DELETE = 'DATASET_DELETE';
-const DATASET_DELETE_VALUES = 'DATASET_DELETE_VALUES';
+const DATASET_DELETE_SELECTED_VALUES = 'DATASET_DELETE_SELECTED_VALUES';
 const DATASET_DELETE_ALL_VALUES= 'DATASET_DELETE_ALL_VALUES';
 
 class DataSetValueActions extends Component {
@@ -29,14 +29,41 @@ class DataSetValueActions extends Component {
       case DATASET_ENABLE:
         this.props.dataSetActions.updateDataSet(dataSetId, {status: app.STATION_ENABLED});
         break;
-      case DATASET_DELETE:
-        this.props.dataSetActions.deleteDataSet({id: dataSetId});
-        break;
-      case DATASET_DELETE_VALUES:
-        this.props.dataSetActions.deleteSelectedDataSetValues('id');
+      case DATASET_DELETE_SELECTED_VALUES:
+        Modal.confirm({
+          title: 'Delete selected values',
+          content: 'Are you sure want delete selected values in this data set?',
+          okText: 'Yes',
+          okType: 'danger',
+          cancelText: 'No',
+          onOk: () => {
+            this.props.dataSetActions.deleteSelectedDataSetValues('id')
+          },
+        });
         break;
       case DATASET_DELETE_ALL_VALUES:
-        this.props.dataSetActions.deleteDataSetValue({dataSetId});
+        Modal.confirm({
+          title: 'Delete all values',
+          content: 'Are you sure want delete all values in this data set?',
+          okText: 'Yes',
+          okType: 'danger',
+          cancelText: 'No',
+          onOk: () => {
+            this.props.dataSetActions.deleteDataSetValue({dataSetId})
+          },
+        });
+        break;
+      case DATASET_DELETE:
+        Modal.confirm({
+          title: 'Delete data set',
+          content: 'Are you sure want delete this data set?',
+          okText: 'Yes',
+          okType: 'danger',
+          cancelText: 'No',
+          onOk: () => {
+            this.props.dataSetActions.deleteDataSet({id: dataSetId})
+          },
+        });
         break;
       default:
         break;
@@ -60,7 +87,7 @@ class DataSetValueActions extends Component {
           :
           <Menu.Item key={DATASET_ENABLE}><Icon type="check-square-o" /> Enable</Menu.Item>
         }
-        <Menu.Item key={DATASET_DELETE_VALUES}><Icon type="bars" /> Delete selected values</Menu.Item>
+        <Menu.Item key={DATASET_DELETE_SELECTED_VALUES}><Icon type="bars" /> Delete selected values</Menu.Item>
         <Menu.Item key={DATASET_DELETE_ALL_VALUES}><Icon type="table" /> Delete all values</Menu.Item>
         <Menu.Item key={DATASET_DELETE}><Icon type="delete" /> Delete data set</Menu.Item>
       </Menu>
